refactor: keep sweep defaults in infx.config
将扫描默认路径保留在独立的 infx.config 模块中，维持旧导入兼容性并保持包入口精简。
