+++ infx/config.py
@@ -1,0 +1,5 @@
+"""Default repository paths used by sweep planning."""
+
+MASTER_CONFIGS = ["configs/amd-master.yaml", "configs/nvidia-master.yaml"]
+RUNNER_CONFIG = "configs/runners.yaml"
+GENERATE_SWEEPS_PY_SCRIPT = "utils/matrix_logic/generate_sweep_configs.py"
